test(ui): split app sidebar suite
