+++ ui/src/test-helpers/app-sidebar-cases/basics.ts
@@ -1,0 +1,328 @@
+import { describe, expect, it, vi } from "vitest";
+import type { GatewayBrowserClient } from "../../api/gateway.ts";
+import type { AgentsListResult } from "../../api/types.ts";
+import {
+  createGateway,
+  createGatewayHarness,
+  createSessions,
+  createSessionsHarness,
+  mountSidebar,
+  TWO_AGENTS,
+} from "../app-sidebar.ts";
+import "../../components/app-sidebar.ts";
+
+describe("AppSidebar update card wiring", () => {
+  it("shows OpenClaw in the default pinned routes", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const { sidebar } = await mountSidebar(gateway, createSessions("main", ["agent:main:main"]));
+
+    const link = sidebar.querySelector<HTMLAnchorElement>('.nav-item[href="/custodian"]');
+    expect(link?.textContent?.trim()).toBe("OpenClaw");
+  });
+
+  it("renders the update card in the footer after the attention slot and forwards its action", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const { sidebar } = await mountSidebar(gateway, createSessions("main", ["agent:main:main"]));
+    const onUpdate = vi.fn();
+    sidebar.updateAvailable = {
+      currentVersion: "1.0.0",
+      latestVersion: "2.0.0",
+      channel: "stable",
+    };
+    sidebar.onUpdate = onUpdate;
+    await sidebar.updateComplete;
+
+    const footer = sidebar.querySelector(".sidebar-shell__footer");
+    // Attention chips (when present) stack above the update card.
+    expect(footer?.firstElementChild?.localName).toBe("openclaw-sidebar-attention");
+    const card = footer?.querySelector("openclaw-sidebar-update-card");
+    expect(card).not.toBeNull();
+    card?.querySelector<HTMLButtonElement>(".sidebar-update-card__action")?.click();
+    expect(onUpdate).toHaveBeenCalledOnce();
+  });
+});
+
+describe("AppSidebar brand actions", () => {
+  it("starts a session for the active agent from the Threads header", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const agentsList = {
+      defaultId: "main",
+      mainKey: "main",
+      scope: "agent",
+      agents: [{ id: "main" }, { id: "research" }],
+    } as AgentsListResult;
+    const { sidebar } = await mountSidebar(
+      gateway,
+      createSessions("research", ["agent:research:main"]),
+      "panel",
+      agentsList,
+    );
+    const onOpenNewSession = vi.fn();
+    sidebar.connected = true;
+    sidebar.onOpenNewSession = onOpenNewSession;
+    await sidebar.updateComplete;
+
+    // The new-session affordance moved off the brand row onto Threads.
+    expect(sidebar.querySelector(".sidebar-brand .sidebar-new-session")).toBeNull();
+    expect(sidebar.querySelector(".sidebar-recent-sessions__head--root")).toBeNull();
+    const button = sidebar.querySelector<HTMLButtonElement>(
+      '[data-session-section="ungrouped"] .sidebar-new-session',
+    );
+    expect(button?.getAttribute("aria-label")).toBe("New session");
+    expect(button?.disabled).toBe(false);
+
+    button?.click();
+    expect(onOpenNewSession).toHaveBeenCalledExactlyOnceWith("research");
+  });
+});
+
+describe("AppSidebar agent chip", () => {
+  it("resumes the newest session when the menu switches to an agent with cached rows", async () => {
+    const gatewayHarness = createGatewayHarness({} as GatewayBrowserClient);
+    const setSessionKey = vi.fn();
+    (gatewayHarness.gateway as { setSessionKey: (key: string) => void }).setSessionKey =
+      setSessionKey;
+    const { sidebar } = await mountSidebar(
+      gatewayHarness.gateway,
+      createSessions("main", ["agent:main:main", "agent:main:task"]),
+      "panel",
+      TWO_AGENTS,
+    );
+    const onNavigate = vi.fn();
+    sidebar.connected = true;
+    sidebar.onNavigate = onNavigate;
+    await sidebar.updateComplete;
+
+    sidebar.querySelector<HTMLButtonElement>(".sidebar-agent-card__main")?.click();
+    await sidebar.updateComplete;
+    const rows = [
+      ...sidebar.querySelectorAll<HTMLElement>(
+        ".sidebar-agent-menu wa-dropdown-item.sidebar-agent-menu__agent-switch",
+      ),
+    ];
+    rows.find((row) => row.textContent?.includes("Molty"))?.click();
+    // createSessionState stamps ascending updatedAt, so the last key is newest.
+    expect(setSessionKey).toHaveBeenCalledWith("agent:main:task");
+    expect(onNavigate).toHaveBeenCalledWith("chat", { search: "?session=agent%3Amain%3Atask" });
+  });
+
+  it("keeps agent ids distinct from utility command values", async () => {
+    const gatewayHarness = createGatewayHarness({} as GatewayBrowserClient);
+    const setSessionKey = vi.fn();
+    (gatewayHarness.gateway as { setSessionKey: (key: string) => void }).setSessionKey =
+      setSessionKey;
+    const agents = {
+      defaultId: "main",
+      mainKey: "main",
+      scope: "agent",
+      agents: [{ id: "main" }, { id: "settings" }],
+    } as AgentsListResult;
+    const { sidebar } = await mountSidebar(
+      gatewayHarness.gateway,
+      createSessions("main", ["agent:main:main"]),
+      "panel",
+      agents,
+    );
+    const onNavigate = vi.fn();
+    sidebar.connected = true;
+    sidebar.onNavigate = onNavigate;
+    await sidebar.updateComplete;
+
+    sidebar.querySelector<HTMLButtonElement>(".sidebar-agent-card__main")?.click();
+    await sidebar.updateComplete;
+    const menu = sidebar.querySelector<HTMLElement>(".sidebar-agent-menu");
+    const settingsAgent = [
+      ...(menu?.querySelectorAll<HTMLElement>('wa-dropdown-item[type="checkbox"]') ?? []),
+    ].find((row) => row.textContent?.includes("settings"));
+    menu?.dispatchEvent(
+      new CustomEvent("wa-select", { detail: { item: settingsAgent }, bubbles: true }),
+    );
+    await sidebar.updateComplete;
+
+    expect(setSessionKey).toHaveBeenCalledWith("agent:settings:main");
+    expect(onNavigate).toHaveBeenCalledWith("chat", {
+      search: "?session=agent%3Asettings%3Amain",
+    });
+    expect(onNavigate).not.toHaveBeenCalledWith("config");
+  });
+
+  it("shows offline in the chip subtitle when disconnected", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const { sidebar } = await mountSidebar(gateway, createSessions("main", ["agent:main:main"]));
+    sidebar.connected = false;
+    await sidebar.updateComplete;
+
+    expect(sidebar.querySelector(".sidebar-agent-card__subtitle")?.textContent?.trim()).toBe(
+      "Offline",
+    );
+  });
+
+  it("shows a working subtitle while the agent has an active run", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const harness = createSessionsHarness("main", ["agent:main:main"]);
+    const { sidebar } = await mountSidebar(gateway, harness.sessions);
+    sidebar.connected = true;
+    harness.publishList({
+      result: {
+        ts: 2,
+        path: "",
+        count: 1,
+        defaults: { modelProvider: null, model: null, contextTokens: null },
+        sessions: [{ key: "agent:main:main", kind: "direct", updatedAt: 5, hasActiveRun: true }],
+      },
+      agentId: "main",
+    });
+    await sidebar.updateComplete;
+
+    expect(sidebar.querySelector(".sidebar-agent-card__subtitle")?.textContent).toContain(
+      "Working",
+    );
+  });
+
+  it("keeps the sessions list flat for the selected agent and flags other-agent unread", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const harness = createSessionsHarness("main", ["agent:main:main"]);
+    const { sidebar, context } = await mountSidebar(gateway, harness.sessions, "panel", TWO_AGENTS);
+    sidebar.connected = true;
+    const defaults = { modelProvider: null, model: null, contextTokens: null };
+    harness.publishList({
+      result: {
+        ts: 2,
+        path: "",
+        count: 1,
+        defaults,
+        sessions: [
+          {
+            key: "agent:research:one",
+            kind: "direct",
+            label: "Research task",
+            updatedAt: 3,
+            unread: true,
+          },
+        ],
+      },
+      agentId: "research",
+    });
+    harness.publishList({
+      result: {
+        ts: 3,
+        path: "",
+        count: 1,
+        defaults,
+        sessions: [{ key: "agent:main:main", kind: "direct", updatedAt: 5 }],
+      },
+      agentId: "main",
+    });
+    await sidebar.updateComplete;
+
+    // No per-agent sections: the card switcher owns agent switching now, and
+    // the main session lives behind the identity card instead of the list.
+    expect(sidebar.querySelector(".sidebar-agent-section")).toBeNull();
+    expect(sidebar.querySelectorAll(".sidebar-recent-session")).toHaveLength(0);
+    expect(sidebar.querySelector(".sidebar-agent-card__menu-unread")).not.toBeNull();
+
+    // Mid-switch (selected agent != loaded result agent) the list renders the
+    // target agent's cached rows instead of flashing empty until refresh.
+    // Chip switch and chat-pane both sync agentSelection with the route.
+    context.agentSelection.state.selectedId = "research";
+    sidebar.sessionKey = "agent:research:one";
+    await sidebar.updateComplete;
+    const rows = [...sidebar.querySelectorAll(".sidebar-recent-session")];
+    expect(rows).toHaveLength(1);
+    expect(rows[0]?.textContent).toContain("Research task");
+  });
+
+  it("routes Home to the main session and marks it active there", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const setSessionKey = vi.fn();
+    (gateway as { setSessionKey: (key: string) => void }).setSessionKey = setSessionKey;
+    const { sidebar } = await mountSidebar(gateway, createSessions("main", ["agent:main:main"]));
+    const navigate = vi.fn();
+    sidebar.onNavigate = navigate;
+    sidebar.connected = true;
+    (sidebar as unknown as { activeRouteId: string }).activeRouteId = "chat";
+    sidebar.sessionKey = "agent:main:main";
+    await sidebar.updateComplete;
+
+    const home = sidebar.querySelector<HTMLAnchorElement>(".nav-item--home");
+    expect(home?.textContent).toContain("Home");
+    expect(home?.getAttribute("aria-current")).toBe("page");
+
+    home?.click();
+    expect(setSessionKey).toHaveBeenCalledWith("agent:main:main");
+    expect(navigate).toHaveBeenCalledWith("chat", { search: "?session=agent%3Amain%3Amain" });
+  });
+
+  it("treats the global key as the main session under global scope", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    const harness = createSessionsHarness("main", ["global"]);
+    const globalAgents = {
+      defaultId: "main",
+      mainKey: "main",
+      scope: "global",
+      agents: [{ id: "main", identity: { name: "Molty" } }],
+    } as AgentsListResult;
+    const { sidebar } = await mountSidebar(gateway, harness.sessions, "panel", globalAgents);
+    harness.publishList({
+      result: {
+        ts: 2,
+        path: "",
+        count: 2,
+        defaults: { modelProvider: null, model: null, contextTokens: null },
+        sessions: [
+          { key: "global", kind: "global", updatedAt: 5, unread: true },
+          { key: "agent:main:side-quest", kind: "direct", label: "Side quest", updatedAt: 4 },
+        ],
+      },
+    });
+    await sidebar.updateComplete;
+
+    // The advertised global main hides behind the Home row instead of
+    // leaking into Threads; ordinary sessions still list, and Home surfaces
+    // the global row's unread state.
+    expect(sidebar.querySelector('[data-session-key="global"]')).toBeNull();
+    expect(sidebar.querySelector('[data-session-key="agent:main:side-quest"]')).not.toBeNull();
+    expect(sidebar.querySelector(".nav-item--home .session-unread-dot")).not.toBeNull();
+  });
+
+  it("promotes main-session children to top-level threads, including alias parent keys", async () => {
+    const gateway = createGateway({} as GatewayBrowserClient);
+    // The gateway row uses the unprefixed "main" alias; children index under
+    // that literal key, so promotion must follow the row's key, not only the
+    // synthesized agent:main:main form.
+    const harness = createSessionsHarness("main", ["main"]);
+    const { sidebar } = await mountSidebar(gateway, harness.sessions);
+    harness.publishList({
+      result: {
+        ts: 2,
+        path: "",
+        count: 2,
+        defaults: { modelProvider: null, model: null, contextTokens: null },
+        sessions: [
+          {
+            key: "main",
+            kind: "direct",
+            updatedAt: 5,
+            childSessions: ["agent:main:subagent:thread-a"],
+          },
+          {
+            key: "agent:main:subagent:thread-a",
+            spawnedBy: "main",
+            kind: "direct",
+            label: "Spawned thread",
+            updatedAt: 4,
+          },
+        ],
+      },
+    });
+    await sidebar.updateComplete;
+
+    // The main row hides behind the identity card; its child surfaces as a
+    // top-level (non-child) thread row.
+    expect(sidebar.querySelector('[data-session-key="main"]')).toBeNull();
+    const promoted = sidebar.querySelector('[data-session-key="agent:main:subagent:thread-a"]');
+    expect(promoted).not.toBeNull();
+    expect(promoted?.classList.contains("sidebar-recent-session--child")).toBe(false);
+    expect(promoted?.textContent).toContain("Spawned thread");
+  });
+});
