+++ ui/src/test-helpers/app-sidebar-suite.ts
@@ -1,0 +1,3 @@
+import { setupSidebarTest } from "./app-sidebar.ts";
+
+setupSidebarTest();
